+++ app/src/main/java/cn/sgr/zmr/com/sgr/Common/MainActivity.java
@@ -14,12 +14,7 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import cn.sgr.zmr.com.sgr.Base.MyApplication;
-import cn.sgr.zmr.com.sgr.Common.Login.LoginActivity;
-import cn.sgr.zmr.com.sgr.Common.Model.UserInfo;
 import cn.sgr.zmr.com.sgr.Modules.Health.HealhFragment;
-import cn.sgr.zmr.com.sgr.Modules.Home.HomeFragment;
-import cn.sgr.zmr.com.sgr.Modules.Home.HomePresenter;
-import cn.sgr.zmr.com.sgr.Modules.Home.HomeTestContract;
 import cn.sgr.zmr.com.sgr.Modules.Home.HomeTestFragment;
 import cn.sgr.zmr.com.sgr.Modules.Home.HomeTestPresenter;
 import cn.sgr.zmr.com.sgr.Modules.Messages.MessageFragment;
