+++ app/src/main/java/cn/sgr/zmr/com/sgr/Modules/Home/HomeTestContract.java
@@ -2,6 +2,7 @@ package cn.sgr.zmr.com.sgr.Modules.Home;
 
 import android.app.Activity;
 import android.app.FragmentManager;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.support.annotation.ColorRes;
 import android.view.View.OnClickListener;
@@ -26,99 +27,13 @@ public class HomeTestContract {
 
         void cancelProgressDialog();//隐藏进度条
 
-        void clearData();//清楚数据
-
-        void setDate(Baby baby);//设置宝宝数据
-
-//        void initBlue();//初始化蓝牙
-
-        // TODO
-        void setBoundedBabyData(Baby baby);
-
-        // TODO 参数待改，可能有变化
-        void showChooseBabyDialog(List<Baby> babyList, String positiveBtnText, String negativeBtnText);
-
-        // TODO
-        /**
-         * 显示消息对话框
-         * @param title
-         *          标题
-         * @param optimisticBtnText
-         *          右边确定按钮文本
-         * @param negativeBtnText
-         *          左边取消按钮文本
-         * @param positiveBtnListener
-         *          右边确定按钮监听器
-         * @param negativeBtnListener
-         *          左边取消按钮监听器
-         */
-        void showMsgDialog(String title, String message, String optimisticBtnText, String negativeBtnText,
-                           DialogInterface.OnClickListener positiveBtnListener,
-                           DialogInterface.OnClickListener negativeBtnListener);
-
-        void setTemperatureDisplay(int temperature, @ColorRes int themeColor);
-
-        void setTemperatureStateDisplay(String temperatureState, @ColorRes int themeColor);
-
-        void setBatteryInfoDisplay(String batteryInfo, String batteryNum, @ColorRes int themeColor);
-
-        void setCircularBarStatus(int startNum, int endNum, int duration, @ColorRes int themeColor);
-
-        void setTopBarLeftText(String text);
-
-        void setHeight(int height);
-
-        void setWeight(int weight);
-
-        void setBirthday(Date date);
+        void showUserInfo(Context context, String nickName, String phone, String avatar);//显示用户信息
 
         boolean isActive();//目的是为了解决内存泄漏
     }
 
     interface Presenter extends BasePresenter {
-        /**
-         * 初始化蓝牙
-         * @param activity
-         */
-        void initBlueTooth(Activity activity);//初始化蓝牙
 
-        /**
-         * 连接设备
-         * @param activity
-         */
-        void linkDevice(Activity activity);//连接设备
-
-        void initDevice(Activity activity);
-
-        /**
-         * 设置要绑定的宝宝
-         * @param boundedBaby
-         */
-        void setBoundedBaby(Baby boundedBaby);
-
-        /**
-         * 绑定宝宝
-         */
-        void boundBaby();//绑定宝宝
-
-        /**
-         *
-         */
-        void boundBabyandDevice(String DeviceName);
-
-        /**
-         * 新建宝宝
-         */
-        void createBaby();// 新建宝宝
-
-        /**
-         * 发送指令
-         * @param activity
-         * @param deviceAddress
-         */
-        void sendCmd(Activity activity, String deviceAddress);// 发送指令
-
-        void closeDevice();
 
 
     }
