+++ app/src/main/java/cn/sgr/zmr/com/sgr/Modules/Home/Module/Device/DeviceFragment.java
@@ -13,7 +13,8 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 import cn.sgr.zmr.com.sgr.Base.BaseFragment;
 import cn.sgr.zmr.com.sgr.Common.Model.Setting;
-import cn.sgr.zmr.com.sgr.Modules.Home.HomeFragment;
+import cn.sgr.zmr.com.sgr.Modules.Home.Device.HomeDeviceFragment;
+import cn.sgr.zmr.com.sgr.Modules.Home.Device.HomeDevicePresenter;
 import cn.sgr.zmr.com.sgr.R;
 
 /**
@@ -55,7 +56,7 @@ public class DeviceFragment extends BaseFragment implements DeviceContract.View{
     private void initView() {
         top_view_title.setText(getResources().getString(R.string.set_decive ));
         top_view_back.setVisibility(View.VISIBLE);
-        if(HomeFragment.isConnState){
+        if(HomeDevicePresenter.isConnState){
 
         }else{
             text_quantity.setText(Setting.getInstance(getActivity()).getBattery());
