+++ app/src/main/res/drawable/green_select.xml
@@ -2,7 +2,7 @@
 <selector xmlns:android="http://schemas.android.com/apk/res/android" >
     <item android:state_pressed="true">
         <shape android:shape="rectangle">
-            <solid android:color="#4ea04e"/>
+            <solid android:color="@color/them_bg_deep"/>
         </shape>
     </item>
     <item android:state_focused="true"
