+++ app/src/main/res/layout/chart_fragment.xml
@@ -7,11 +7,11 @@
 
 
 
-    <com.sothree.slidinguppanel.SlidingUpPanelLayout
+<!--    <com.sothree.slidinguppanel.SlidingUpPanelLayout
         android:layout_margin="5dp"
         android:id="@+id/sliding_layout"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent">-->
 
         <LinearLayout
             android:layout_marginTop="10dp"
@@ -112,7 +112,7 @@
 
 
 
-
+<!--
         <LinearLayout
             android:layout_marginTop="10dp"
 
@@ -157,5 +157,6 @@
         </LinearLayout>
 
     </com.sothree.slidinguppanel.SlidingUpPanelLayout>
+    -->
 
 </LinearLayout>
