+++ app/src/main/res/values/colors.xml
@@ -9,7 +9,7 @@
     <color name="my_bg">#ebebeb</color>
 
     <color name="them_bg">#ea4c89</color>
-    <color name="them_bg_deep">#6cb950</color>
+    <color name="them_bg_deep">#ca4c8a</color>
     <color name="gery">#CCCCCC</color>
 
     <color name="them_text">#888888</color>
